+++ src/pages/ManageBlog.jsx
@@ -157,7 +157,7 @@ const ManageBlog = () => {
   };
   const handleSubmitForm = async () => {
     const requiredFields = ["Name", "Category", "Description"];
-    const emptyRequiredFields = requiredFields.filter((field) => !data[field]);
+    const emptyRequiredFields = requiredFields.filter( (field) => !data[field] || !data[field].trim());
     if (emptyRequiredFields.length > 0 || selectedTags.length === 0) {
       validationAlert("Please fill in all required fields");
       return;
@@ -244,23 +244,23 @@ const ManageBlog = () => {
 
       if (result.isConfirmed) {
         try {
-          const res = await axios.request({
-            method: "PUT",
-            maxBodyLength: Infinity,
-            url: `${Bunny_Storage_URL}/Blogs/${filename}`,
-            headers: {
-              "Content-Type": "image/jpeg",
-              AccessKey: Bunny_Storage_Access_Key,
-            },
-            data: uploadedImg,
-          });
+          const response = await axios.patch(
+            `${BASE_URL}blogs/${data.Id}`,
+            UpdateObj
+          );
 
-          if (res.data.HttpCode === 201) {
-            const response = await axios.patch(
-              `${BASE_URL}blogs/${data.Id}`,
-              UpdateObj
-            );
-            if (response.data.status) {
+          if (response.data.status && uploadedImg !== "") {
+            const res = await axios.request({
+              method: "PUT",
+              maxBodyLength: Infinity,
+              url: `${Bunny_Storage_URL}/Blogs/${filename}`,
+              headers: {
+                "Content-Type": "image/jpeg",
+                AccessKey: Bunny_Storage_Access_Key,
+              },
+              data: uploadedImg,
+            });
+            if (res.data.HttpCode === 201) {
               if (uploadedImg !== "") {
                 await axios.request({
                   method: "DELETE",
@@ -288,8 +288,20 @@ const ManageBlog = () => {
               throw new Error("Failed to Update Data");
             }
           } else {
+            // setLoaderOpen(false);
+            // throw new Error("Failed to Upload Image");
             setLoaderOpen(false);
-            throw new Error("Failed to Upload Image");
+              Swal.fire({
+                position: "center",
+                icon: "success",
+                title: "Data Updated Successfully",
+                toast: true,
+                showConfirmButton: false,
+                timer: 1500,
+              });
+              handleClose();
+              getAllImgList();
+              setUploadedImg("");
           }
         } catch (error) {
           setLoaderOpen(false);
@@ -735,10 +747,14 @@ const ManageBlog = () => {
       </Grid>
 
       <Grid container spacing={3} width="100%" pt={5}>
-        <Grid item xs={12}>
+        <Grid
+          item
+          xs={12}
+          style={{ display: "flex", justifyContent: "center" }}
+        >
           <Pagination
-            count={Math.ceil(19 / 8)}
-            color="primary"
+            count={Math.ceil(imgData.length / 8)}
+            color="secondary"
             page={page}
             onChange={handlePageChange}
           />
