+++ src/pages/ManageSchedule/Diet.jsx
@@ -113,7 +113,6 @@ const ManageDiet = () => {
       setUploadedImg("");
     }
   };
-  
 
   const handleClose = () => {
     setOn(false);
@@ -146,7 +145,10 @@ const ManageDiet = () => {
 
   const handleSubmitForm = async () => {
     const requiredFields = ["Name", "Description"];
-    const emptyRequiredFields = requiredFields.filter((field) => !data[field].trim());
+    const emptyRequiredFields = requiredFields.filter(
+      // (field) => !data[field]||"".trim()
+      (field) => !data[field] || !data[field].trim()
+    );
     if (emptyRequiredFields.length > 0 || selectedTags.length === 0) {
       validationAlert("Please fill in all required fields");
       return;
@@ -231,23 +233,23 @@ const ManageDiet = () => {
 
       if (result.isConfirmed) {
         try {
-          const res = await axios.request({
-            method: "PUT",
-            maxBodyLength: Infinity,
-            url: `${Bunny_Storage_URL}/Schedule/Diet/${filename}`,
-            headers: {
-              "Content-Type": "image/jpeg",
-              AccessKey: Bunny_Storage_Access_Key,
-            },
-            data: uploadedImg,
-          });
+          const response = await axios.patch(
+            `${BASE_URL}diet/${data.Id}`,
+            UpdateObj
+          );
 
-          if (res.data.HttpCode === 201) {
-            const response = await axios.patch(
-              `${BASE_URL}diet/${data.Id}`,
-              UpdateObj
-            );
-            if (response.data.status) {
+          if (response.data.status && uploadedImg !== "") {
+            const res = await axios.request({
+              method: "PUT",
+              maxBodyLength: Infinity,
+              url: `${Bunny_Storage_URL}/Schedule/Diet/${filename}`,
+              headers: {
+                "Content-Type": "image/jpeg",
+                AccessKey: Bunny_Storage_Access_Key,
+              },
+              data: uploadedImg,
+            });
+            if (res.data.HttpCode === 201) {
               if (uploadedImg !== "") {
                 await axios.request({
                   method: "DELETE",
@@ -276,7 +278,17 @@ const ManageDiet = () => {
             }
           } else {
             setLoaderOpen(false);
-            throw new Error("Failed to Upload Image");
+            Swal.fire({
+              position: "center",
+              icon: "success",
+              title: "Data Updated Successfully",
+              toast: true,
+              showConfirmButton: false,
+              timer: 1500,
+            });
+            handleClose();
+            getAllImgList();
+            setUploadedImg("");
           }
         } catch (error) {
           setLoaderOpen(false);
@@ -297,7 +309,7 @@ const ManageDiet = () => {
   const getAllImgList = () => {
     axios.get(`${BASE_URL}diet/`).then((response) => {
       setImgData(response.data.values.flat());
-    });
+    });   
   };
 
   const getTagData = () => {
@@ -543,7 +555,7 @@ const ManageDiet = () => {
                 tabIndex={-1}
                 startIcon={<CloudUploadIcon />}
                 sx={{
-                  backgroundColor: "#8F00FF", 
+                  backgroundColor: "#8F00FF",
                   py: 1.5,
                   "&:hover": {
                     backgroundColor: "#3B444B",
@@ -728,10 +740,14 @@ const ManageDiet = () => {
       </Grid>
 
       <Grid container spacing={3} width="100%" pt={5}>
-        <Grid item xs={12}>
+        <Grid
+          item
+          xs={12}
+          style={{ display: "flex", justifyContent: "center" }}
+        >
           <Pagination
-            count={Math.ceil(19 / 8)}
-            color="primary"
+            count={Math.ceil(imgData.length / 8)}
+            color="secondary"
             page={page}
             onChange={handlePageChange}
           />
